refactor: move tokio benchmark together
